Feature / Comment Analytics : Analytics Service (Listener)
Cross-review fixes:

Resolves Conflicts;
Fixes Broken tests in master;
Refactors code in master;

+++ src/main/java/faang/school/analytics/config/RetryProperties.java
@@ -18,7 +18,7 @@ public class RetryProperties {
     @Min(1000)
     private long initialDelay;
     @Min(1)
-    private int multiplier;
+    private double multiplier;
     @Min(1000)
     private long maxDelay;
 }
